# sdk/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ClosePositionDto": {
        "properties": {
          "exit_date": {
            "description": "Exit date (ISO date string)",
            "example": "2026-05-10",
            "type": "string"
          },
          "exit_price": {
            "description": "Exit price in USD",
            "example": 2500,
            "type": "number"
          },
          "exit_reason": {
            "description": "Exit reason",
            "example": "take_profit",
            "type": "string"
          },
          "pnl_percent": {
            "description": "Final P&L percent",
            "type": "number"
          },
          "pnl_usd": {
            "description": "Final P&L in USD",
            "type": "number"
          }
        },
        "required": ["exit_price"],
        "type": "object"
      },
...
      "CreatePositionDto": {
        "properties": {
          "address": {
            "description": "Token contract address",
            "example": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
            "type": "string"
          },
          "chain": {
            "description": "Chain identifier",
            "example": "base",
            "type": "string"
          },
          "entry_date": {
            "description": "Entry date (ISO date string)",
            "example": "2026-05-10",
            "type": "string"
          },
          "entry_price": {
            "description": "Entry price in USD",
            "example": 2000,
            "type": "number"
          },
          "mode": {
            "description": "Portfolio mode (default: real)",
            "enum": ["real", "paper"],
            "type": "string"
          },
          "name": {
            "description": "Token name",
            "example": "Ethereum",
            "type": "string"
          },
          "narrative": {
            "description": "Narrative tag",
            "example": "defi",
            "type": "string"
          },
          "notes": {
            "description": "Notes",
            "type": "string"
          },
          "quantity": {
            "description": "Token quantity",
            "example": 0.5,
            "type": "number"
          },
          "stop_loss": {
            "description": "Stop-loss price in USD",
            "example": 1600,
            "type": "number"
          },
          "symbol": {
            "description": "Token symbol",
            "example": "ETH",
            "type": "string"
          },
          "take_profit_levels": {
            "description": "Take-profit price levels (JSON array of numbers)",
            "example": [2500, 3000, 4000],
            "items": {
              "type": "number"
            },
            "type": "array"
          },
          "tier": {
            "description": "Position tier",
            "enum": ["base", "conviction", "moonshot"],
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "address",
          "chain",
          "tier",
          "entry_price",
          "quantity",
          "stop_loss",
          "take_profit_levels"
        ],
        "type": "object"
      },
...
      "UpdatePositionDto": {
        "properties": {
          "current_price": {
            "description": "Updated current price",
            "type": "number"
          },
          "last_synced_at": {
            "description": "Timestamp of last on-chain sync",
            "type": "string"
          },
          "max_price_since_entry": {
            "description": "Max price observed since entry (for trailing stop)",
            "type": "number"
          },
          "narrative": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "onchain_balance": {
            "description": "On-chain balance (synced from chain)",
            "type": "number"
          },
          "quantity": {
            "description": "Updated quantity",
            "type": "number"
          },
          "status": {
            "enum": ["open", "partial_exit", "closed", "pending_analysis", "draft", "pending_exit"],
            "type": "string"
          },
          "stop_loss": {
            "description": "Updated stop-loss price",
            "type": "number"
          },
          "take_profit_levels": {
            "description": "Updated take-profit levels (array of numbers)",
            "items": {
              "type": "number"
            },
            "type": "array"
          },
          "tp_levels_hit": {
            "description": "TP levels already hit (JSON array of booleans/indices)",
            "items": {
              "type": "number"
            },
            "type": "array"
          },
          "trailing_stop_pct": {
            "description": "Trailing stop percentage (0–100)",
            "type": "number"
          },
          "value_usd": {
            "description": "Updated value in USD",
            "type": "number"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "CryptoClaw API — auto-generated from controllers + DTOs",
    "title": "CryptoClaw",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/positions": {
      "get": {
        "operationId": "PositionsController_list",
        "parameters": [
          {
            "description": "Filter by position status",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": ["open", "partial_exit", "closed", "pending_analysis", "draft", "pending_exit"],
              "type": "string"
            }
          },
          {
            "description": "Portfolio mode (default: real)",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "enum": ["real", "paper"],
              "type": "string"
            }
          },
          {
            "description": "Filter by token symbol (case-insensitive)",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by chain",
            "in": "query",
            "name": "chain",
            "required": false,
            "schema": {
              "example": "base",
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results (default 50, max 200)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Cursor for pagination (last position id)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of positions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List positions",
        "tags": ["positions"]
      },
      "post": {
        "operationId": "PositionsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePositionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Position created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new position",
        "tags": ["positions"]
      }
    },
    "/v1/positions/{id}": {
      "delete": {
        "operationId": "PositionsController_delete",
        "parameters": [
          {
            "description": "Position ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Position deleted"
          },
          "404": {
            "description": "Position not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a position",
        "tags": ["positions"]
      },
      "get": {
        "operationId": "PositionsController_getById",
        "parameters": [
          {
            "description": "Position ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Position found"
          },
          "404": {
            "description": "Position not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a position by ID",
        "tags": ["positions"]
      },
      "patch": {
        "operationId": "PositionsController_update",
        "parameters": [
          {
            "description": "Position ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Position updated"
          },
          "404": {
            "description": "Position not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a position",
        "tags": ["positions"]
      }
    },
    "/v1/positions/{id}/close": {
      "post": {
        "operationId": "PositionsController_close",
        "parameters": [
          {
            "description": "Position ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosePositionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Position closed"
          },
          "404": {
            "description": "Position not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close a position",
        "tags": ["positions"]
      }
    },
...
  },
  "servers": [],
...
}